requirements.txt
gradio==5.23.3
gradio_client==1.8.0
pandas==2.2.3
openpyxl==3.0.10
bs4==0.0.2
lxml==5.3.0
tiktoken==0.8.0
colorama==0.4.6
peewee==3.17.8
fontTools==4.55.3
xlwings==0.33.11
numpy
pdfminer-six==20250416
pymupdf<1.25.3
opencv-python-headless
onnx
onnxruntime
tqdm
tenacity
openai>=1.0.0